remove object navigation flat object method
