+++ packages/twenty-server/src/engine/metadata-modules/metadata-side-effect/handlers/object-metadata/services/__tests__/object-navigation-command-on-create-side-effect-handler.service.spec.ts
@@ -207,160 +207,6 @@ describe('ObjectNavigationCommandOnCreateSideEffectHandlerService', () => {
     expect(result.status).toBe('noop');
   });
 
-  it('noops when a synced engine-owned command already holds the derived identifier', () => {
-    const result = handler.buildSideEffects(
-      buildArgs({
-        syncedFlatCommandMenuItems: [
-          {
-            universalIdentifier: DERIVED_UNIVERSAL_IDENTIFIER,
-            position: 3,
-            engineComponentKey: EngineComponentKey.NAVIGATION,
-            payload: { objectMetadataItemId: OBJECT_ID },
-            isSystemSideEffect: true,
-          } as never,
-        ],
-      }),
-    );
-
-    expect(result.status).toBe('noop');
-  });
-
-  it('still emits when a caller row holds the derived identifier, leaving the squat to the collision detector', () => {
-    const result = handler.buildSideEffects(
-      buildArgs({
-        syncedFlatCommandMenuItems: [
-          {
-            universalIdentifier: DERIVED_UNIVERSAL_IDENTIFIER,
-            position: 3,
-            engineComponentKey: EngineComponentKey.NAVIGATION,
-            payload: { objectMetadataItemId: OBJECT_ID },
-            isSystemSideEffect: false,
-          } as never,
-        ],
-      }),
-    );
-
-    expect(result.status).toBe('success');
-  });
-
-  it('noops when the engine already emitted the command earlier in the same batch', () => {
-    const result = handler.buildSideEffects(
-      buildArgs({
-        pendingFlatCommandMenuItems: [
-          {
-            universalIdentifier: DERIVED_UNIVERSAL_IDENTIFIER,
-            engineComponentKey: EngineComponentKey.NAVIGATION,
-            payload: { objectMetadataItemId: OBJECT_ID },
-            isSystemSideEffect: true,
-          },
-        ],
-      }),
-    );
-
-    expect(result.status).toBe('noop');
-  });
-
-  it('still provisions when a caller-authored row targets the object, so the engine records the collision', () => {
-    const result = expectSuccess(
-      handler.buildSideEffects(
-        buildArgs({
-          pendingFlatCommandMenuItems: [
-            {
-              universalIdentifier: 'manifest-authored-identifier',
-              engineComponentKey: EngineComponentKey.NAVIGATION,
-              payload: { objectMetadataItemId: OBJECT_ID },
-              isSystemSideEffect: false,
-            },
-          ],
-        }),
-      ),
-    );
-
-    expect(
-      Object.keys(result.operations.commandMenuItem?.flatEntityToCreate ?? {}),
-    ).toEqual([DERIVED_UNIVERSAL_IDENTIFIER]);
-  });
-
-  it('still provisions when a caller-authored row squats the derived identifier', () => {
-    const result = expectSuccess(
-      handler.buildSideEffects(
-        buildArgs({
-          pendingFlatCommandMenuItems: [
-            {
-              universalIdentifier: DERIVED_UNIVERSAL_IDENTIFIER,
-              engineComponentKey: EngineComponentKey.NAVIGATION,
-              payload: { objectMetadataItemId: OBJECT_ID },
-              isSystemSideEffect: false,
-            },
-          ],
-        }),
-      ),
-    );
-
-    const [navigationCommand] = Object.values(
-      result.operations.commandMenuItem?.flatEntityToCreate ?? {},
-    ) as unknown as [
-      { universalIdentifier: string; isSystemSideEffect: boolean },
-    ];
-
-    expect(navigationCommand.universalIdentifier).toBe(
-      DERIVED_UNIVERSAL_IDENTIFIER,
-    );
-    expect(navigationCommand.isSystemSideEffect).toBe(true);
-  });
-
-  it('still provisions when a synced caller-authored row sits in the object aggregator', () => {
-    const result = expectSuccess(
-      handler.buildSideEffects(
-        buildArgs({
-          flatObjectMetadata: buildFlatObjectMetadata({
-            commandMenuItemUniversalIdentifiers: ['caller-authored-identifier'],
-          }),
-          syncedFlatCommandMenuItems: [
-            {
-              universalIdentifier: 'caller-authored-identifier',
-              position: 3,
-              engineComponentKey: EngineComponentKey.NAVIGATION,
-              payload: { objectMetadataItemId: OBJECT_ID },
-              isSystemSideEffect: false,
-            } as never,
-          ],
-        }),
-      ),
-    );
-
-    expect(
-      Object.keys(result.operations.commandMenuItem?.flatEntityToCreate ?? {}),
-    ).toEqual([DERIVED_UNIVERSAL_IDENTIFIER]);
-  });
-
-  it('does not treat commands outside the object aggregator as duplicates', () => {
-    const result = expectSuccess(
-      handler.buildSideEffects(
-        buildArgs({
-          syncedFlatCommandMenuItems: [
-            {
-              universalIdentifier: 'go-to-settings-identifier',
-              position: 4,
-              engineComponentKey: EngineComponentKey.NAVIGATION,
-              payload: { path: '/settings/profile' },
-            } as never,
-            {
-              universalIdentifier: 'other-object-navigation-identifier',
-              position: 5,
-              engineComponentKey: EngineComponentKey.NAVIGATION,
-              payload: { objectMetadataItemId: OTHER_OBJECT_ID },
-            } as never,
-          ],
-        }),
-      ),
-    );
-
-    expect(
-      Object.keys(result.operations.commandMenuItem?.flatEntityToCreate ?? {}),
-    ).toEqual([DERIVED_UNIVERSAL_IDENTIFIER]);
-  });
-
   it('appends after the synced maximum position', () => {
     const result = expectSuccess(
       handler.buildSideEffects(
